FIX: 비콘의 좌표 정보를 정수가 아닌 실수로 저장

+++ backend/jangan-backend/src/main/java/com/ssafy/jangan_backend/beacon/dto/BeaconNotificationDto.java
@@ -11,8 +11,8 @@ import lombok.Setter;
 public class BeaconNotificationDto{
 	private String beaconName;
 	private Integer beaconCode;
-	private Integer coordX;
-	private Integer coordY;
+	private Double coordX;
+	private Double coordY;
 	private Integer floor;
 	private String imageUrl;
 	private Integer isNewFire;
